# internal/repository/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.22.0

package repository

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Translation struct {
	WordID            string `db:"word_id" json:"word_id"`
	WordTranslationID string `db:"word_translation_id" json:"word_translation_id"`
}

type Word struct {
	ID         string      `db:"id" json:"id"`
	Value      string      `db:"value" json:"value"`
	Lang       string      `db:"lang" json:"lang"`
	LangRegion pgtype.Text `db:"lang_region" json:"lang_region"`
}